Repositoryモジュールを追加＆MediaStoreモジュールを追加
