+++ app/build.gradle
@@ -48,7 +48,7 @@ android {
 
 dependencies {
     implementation project(":common")
-    implementation project(":data")
+    implementation project(":data:mediastore")
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
