+++ app/src/main/java/si/f5/mob/photoapp/MainActivity.kt
@@ -3,13 +3,24 @@ package si.f5.mob.photoapp
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import dagger.hilt.android.AndroidEntryPoint
+import si.f5.mob.mediastore.MediaStore
+import timber.log.Timber
+import javax.inject.Inject
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+    @Inject
+    lateinit var mediaStore: si.f5.mob.mediastore.MediaStore
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         setContent {
             MainScreen()
         }
+
+        val imageList = mediaStore.getImages()
+        Timber.d(imageList.toString())
     }
 }
