+++ app/src/main/java/si/f5/mob/photoapp/PhotoApp.kt
@@ -1,0 +1,16 @@
+package si.f5.mob.photoapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class PhotoApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
