+++ app/src/main/java/si/f5/mob/photoapp/di/AppModule.kt
@@ -1,0 +1,19 @@
+package si.f5.mob.photoapp.di
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import si.f5.mob.mediastore.MediaStore
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+    @Singleton
+    @Provides
+    fun provideMediaStore(@ApplicationContext context: Context) =
+        si.f5.mob.mediastore.MediaStore(context)
+}
