+++ data/mediastore/src/main/java/si/f5/mob/mediastore/Image.kt
@@ -1,0 +1,11 @@
+package si.f5.mob.mediastore
+
+import android.net.Uri
+
+/**
+ * MediaStoreから取得する画像のエンティティクラス
+ */
+data class Image(
+    val uri: Uri,
+    val name: String,
+)
