--- data/src/main/java/si/f5/mob/data/mediastore/entity/Image.kt
@@ -1,11 +1,0 @@
-package si.f5.mob.data.mediastore.entity
-
-import android.net.Uri
-
-/**
- * MediaStoreから取得する画像のエンティティクラス
- */
-data class Image(
-    val uri: Uri,
-    val name: String,
-)
