+++ repository/src/main/AndroidManifest.xml
